:lipstick: [2FA] Updated first visit page with 2FA suggestion
