--- src/FormV2/CgsType.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace App\FormV2;
-
-use App\Validator\Constraints\AcceptCGS;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class CgsType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('accept', CheckboxType::class, [
-                'label' => 'accept_terms_of_use',
-                'constraints' => [new AcceptCGS()],
-            ]);
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([]);
-    }
-}
